SLING-3045 - [logback]Warning logs logged for 'No appenders present in context [sling] for logger'
Instead of removing the console logger asscoaited with root logger by default we check now if root logger has any appender associated with it or not. if not then console logger is not removed

Also moved this logic out as a separate listener and ensure that this listener is invoked at the end of config cycle.

git-svn-id: https://svn.apache.org/repos/asf/sling/trunk@1566597 13f79535-47bb-0310-9956-ffa450edef68